Completed CHud Money, HealthBar, ArmorBar and BreathBar Logic

+++ pccontrol/hud.h
@@ -6,7 +6,7 @@ public:
     static void Draw();
     static void DrawCrossHairs();
     static void DrawMoney(int32_t money, uint8_t alpha);
-    static void RenderHealthBar(int32_t playerId, int32_t x, int32_t y);
-    static void RenderArmorBar(int32_t playerId, int32_t x, int32_t y);
-    static void RenderBreathBar(int32_t playerId, int32_t x, int32_t y);
+    static void RenderHealthBar(int32_t playerId, float x, float y);
+    static void RenderArmorBar(int32_t playerId, float x, float y);
+    static void RenderBreathBar(int32_t playerId, float x, float y);
 };
